server/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
asgiref==3.9.1
attrs==25.3.0
ccxt==4.4.71
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.57.0
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycares==4.5.0
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-deriv-api==0.1.6
python-dotenv==1.0.1
pytz==2025.2
reactivex==4.0.4
requests==2.32.3
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websockets==10.3
yarl==1.18.3
